rentiq_cloud/requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0

pinecone>=3.0.0

torch>=2.1.0
torchvision>=0.16.0

PyJWT>=2.8.0
bcrypt>=4.1.2
cryptography>=42.0.0

streamlit>=1.35.0
plotly>=5.20.0

python-dotenv>=1.0.1
